Implement hook for JavaParserAndPropagatorUtils
Since JavaParserAndPropagatorUtils currently does not offer a way to perform an action upon generating a Java-model, a listener interface is implemented and integrated into the class to allow generating additional Java-models.

Added a minimal listener interface IJavaModelParserListener and integrated it into JavaParserAndPropagatorUtils to react on Java-Models being parsed. This is done to avoid modifying JavaParserAndPropagatorUtils, as well as each other class that makes use of it.

The main purpose of IJavaModelParserListener is to provide a way for the generator tests to generate an additional Java-Model file, which will be used by fitests, each time JavaParserAndPropagatorUtils parses and creates a Resource instance filled with Java-Model elements. If JavaParserAndPropagatorUtils is to be left the way it is, this hook concept can be refactored and used for other things too; such as listening for configuration changes, component detection, propagation start and propagation end.