Refactor AbstractCITest hierarchy
Refactored the AbstractCITest hierarchy to isolate test-specific details in settings classes that implement the HasRepoSettings interface. The logic of making use of the concrete settings classes is encapsulated within the AbstractRepoTest class, which is between AbstractCITest and the concrete test classes. This change allows test-specific details to be accessed from multiple test classes that make use of the same repository and also enables modifications to the said details via extending them. That way, one can implement different test classes to test different aspects of the project using the same repository without much duplication.

One can implement the ability to switch between different settings dynamically in the future by making the stored settings instance in AbstractRepoTest modifiable (via a setter for example). Furthermore, the parameters of type Object[] in HasRepoSettings could be replaced with EnumMap<EnumForParameterNames, Object> to make things more intuitive and more managable, as those parameters get progressively larger and more complicated.