Add new test plug-in fitests
Added a new plug-in "cipm.consistency.fitests" to help test FirstInstance by making use of non-propagated Java-Model files (short: models) to perform various operations on them, in order to test the integrity of the generated models from start to finish, and also to check whether the said operations work as intended.

It currently only has methods to validate Resource instances of the models from above and some assertion methods to verify the cross-referencing between Matches and Diffs in Comparison instances. As of now, concrete names of the models are provided as variables, which have to be modified to perform a specific unit-test for a specific file. It is recommended to use models of Teammates, because they are significantly smaller in size than TeaStore models.

The non-propagated Java-Model files (models) have to be supplied to the tests by putting them into the "commit-based-cipm/fi-tests/cipm.consistency.fitests/target/<TestGroup>" directory, where <TestGroup> is group of the tests (TeammatesTest or TeaStoreTest). They can be generated by making use of the generator tests from the last commit, which terminate upon the generation of the model at the said directory. Once models are generated, they are copied into the "tmp/<TestGroup>" directory at the same level as the "target" directory and then get used by the unit tests. This is done to spare having to re-generate the aforementioned models after each unit test, as it can be very time consuming, even with the generator tests terminating as soon as possible. It is required, because some unit tests may have to modify models.