Add tests that generate resources for fitests
Added generator tests that make use of the hook (introduced in a previous commit) to generate an additional Java-Model file each time JavaParserAndPropagatorUtils parses one for propagation purposes. This semi-automates generating resources required by fitests, meaning the unit tests (also referred to as the generator tests) can be used to generate the said resources. They, however, have to be manually enabled/disabled and require the "target" directory to be empty to avoid CheckoutConflictExceptions. These tests can be fully automated in the future by having the tearDown method erase the "target" directory and everything in it.

To isolate the additional Java-Model generation, FITestResourceGenerator is introduced as the concrete implementation of this functionality and AbstractFITestResourceGenerator for its further isolation from the generator test classes. FITestResourceGenerator was heavily isolated, as the generation logic may require major changes in the future, because of it having to throw and catch its internal exception to forcefully skip the propagation part in JavaParserAndPropagatorUtils.